rewolucja w planach automatyzacji teraz działają plus kilka zmian we frontendzie (TODO: uładnić guziki)

+++ springboot/src/main/java/pl/szebi/sterowanie/DeviceManager.java
@@ -30,6 +30,8 @@ public class DeviceManager {
     }
 
     public Boolean removeDevice(Integer deviceId) {
+        if (getDevice(deviceId).isOn())
+            return false;
         return deviceRepo.delete(deviceId);
     }
